backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.8.0
werkzeug==3.0.1
gunicorn==21.1.0
